fix(api): принимать ION-балансы и строкой, и числом
Сервер перевёл балансы на NUMERIC(20,6) и отдаёт девять полей строками
вместо целых. kotlinx.serialization падала на разборе, ошибка глоталась —
пользователь видел «No connection» и не проходил дальше экрана входа.
Ломались регистрация, вход по почте, вход через Google, дашборд, финиш
тапа и экран Nodes.

Поля читаются через ленивые сериализаторы, которым равно, что пришло —
3000 или "3000.000000". Тем же способом закрыты метрики замеров и суммы
транзакций: там ровно тот же риск смены типа.

Зарезервированный бонус расширен до Long по всей цепочке.

+++ app/src/main/java/com/aura/feature/onboarding/data/remote/dto/OnboardingDto.kt
@@ -18,7 +18,7 @@ data class AuthSessionDto(
 data class OnboardingFlagsDto(
     val inviteScreenPassed: Boolean,
     val bonusPopupShown: Boolean,
-    val reservedBonusIon: Int,
+    val reservedBonusIon: Long,
 )
 
 data class BootConfigDto(
